+++ base/char.jl
@@ -64,7 +64,7 @@ function show(io::IO, c::Char)
             return
         end
     end
-    if isprint(c)
+    if Unicode.isprint(c)
         write(io, 0x27, c, 0x27)
     else
         u = UInt32(c)
@@ -81,6 +81,6 @@ end
 function show(io::IO, ::MIME"text/plain", c::Char)
     show(io, c)
     u = UInt32(c)
-    print(io, ": ", isascii(c) ? "ASCII/" : "", "Unicode U+", hex(u, u > 0xffff ? 6 : 4))
-    print(io, " (category ", UTF8proc.category_abbrev(c), ": ", UTF8proc.category_string(c), ")")
+    print(io, ": ", Unicode.isascii(c) ? "ASCII/" : "", "Unicode U+", hex(u, u > 0xffff ? 6 : 4))
+    print(io, " (category ", Unicode.category_abbrev(c), ": ", Unicode.category_string(c), ")")
 end