+++ base/pkg/entry.jl
@@ -89,7 +89,7 @@ function available()
     for (pkg, vers) in all_avail
         any(x->Types.satisfies("julia", VERSION, x[2].requires), vers) && push!(avail, pkg)
     end
-    sort!(avail, by=lowercase)
+    sort!(avail, by=Base.Unicode.lowercase)
 end
 
 function available(pkg::AbstractString)
@@ -572,7 +572,7 @@ end
 
 function warnbanner(msg...; label="[ WARNING ]", prefix="")
     cols = Base.displaysize(STDERR)[2]
-    str = rpad(lpad(label, div(cols+textwidth(label), 2), "="), cols, "=")
+    str = rpad(lpad(label, div(cols+Base.Unicode.textwidth(label), 2), "="), cols, "=")
     warn(prefix="", str)
     println(STDERR)
     warn(prefix=prefix, msg...)