+++ base/random/misc.jl
@@ -436,7 +436,7 @@ Base.convert(::Type{UInt128}, u::UUID) = u.value
 
 let groupings = [1:8; 10:13; 15:18; 20:23; 25:36]
     function Base.convert(::Type{UUID}, s::AbstractString)
-        s = lowercase(s)
+        s = Base.Unicode.lowercase(s)
 
         if !ismatch(r"^[0-9a-f]{8}(?:-[0-9a-f]{4}){3}-[0-9a-f]{12}$", s)
             throw(ArgumentError("Malformed UUID string"))