+++ base/strings/strings.jl
@@ -4,7 +4,6 @@ include("strings/errors.jl")
 include("strings/substring.jl")
 include("strings/basic.jl")
 include("strings/search.jl")
+include("strings/unicode.jl")
 include("strings/util.jl")
 include("strings/io.jl")
-include("strings/utf8proc.jl")
-using .UTF8proc