+++ stdlib/Unicode/test/runtests.jl
@@ -1,25 +1,29 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
+using Test
+using Unicode
+using Unicode: normalize, isassigned
+
 @testset "string normalization" begin
-    # normalize_string (Unicode normalization etc.):
-    @test normalize_string("\u006e\u0303", :NFC) == "\u00f1"
-    @test "\u006e\u0303" == normalize_string("\u00f1", :NFD)
-    @test normalize_string("\ufb00", :NFC) != "ff"
-    @test normalize_string("\ufb00", :NFKC) == "ff"
-    @test normalize_string("\u006e\u0303\ufb00", :NFKC) == "\u00f1"*"ff"
-    @test normalize_string("\u00f1\ufb00", :NFKD) == "\u006e\u0303"*"ff"
-    @test normalize_string("\u006e\u0303", compose=true) == "\u00f1"
-    @test "\u006e\u0303" == normalize_string("\u00f1", decompose=true)
-    @test normalize_string("\u006e\u0303\u00b5",compat=true) == "\u00f1\u03bc"
-    @test normalize_string("Σσς",casefold=true) == "σσσ"
-    @test normalize_string("∕⁄", lump=true) == "//"
-    @test normalize_string("\ua\n\r\r\ua", newline2lf=true) == "\ua\ua\ua\ua"
-    @test normalize_string("\ua\n\r\r\ua", newline2ls=true) == "\u2028\u2028\u2028\u2028"
-    @test normalize_string("\ua\n\r\r\ua", newline2ps=true) == "\u2029\u2029\u2029\u2029"
-    @test normalize_string("\u00f1", stripmark=true) == "n"
-    @test isempty(normalize_string("\u00ad", stripignore=true))
-    @test normalize_string("\t\r", stripcc=true) == "  "
-    @test normalize_string("\t\r", stripcc=true, newline2ls=true) == " \u2028"
+    # normalize (Unicode normalization etc.):
+    @test normalize("\u006e\u0303", :NFC) == "\u00f1"
+    @test "\u006e\u0303" == normalize("\u00f1", :NFD)
+    @test normalize("\ufb00", :NFC) != "ff"
+    @test normalize("\ufb00", :NFKC) == "ff"
+    @test normalize("\u006e\u0303\ufb00", :NFKC) == "\u00f1"*"ff"
+    @test normalize("\u00f1\ufb00", :NFKD) == "\u006e\u0303"*"ff"
+    @test normalize("\u006e\u0303", compose=true) == "\u00f1"
+    @test "\u006e\u0303" == normalize("\u00f1", decompose=true)
+    @test normalize("\u006e\u0303\u00b5",compat=true) == "\u00f1\u03bc"
+    @test normalize("Σσς",casefold=true) == "σσσ"
+    @test normalize("∕⁄", lump=true) == "//"
+    @test normalize("\ua\n\r\r\ua", newline2lf=true) == "\ua\ua\ua\ua"
+    @test normalize("\ua\n\r\r\ua", newline2ls=true) == "\u2028\u2028\u2028\u2028"
+    @test normalize("\ua\n\r\r\ua", newline2ps=true) == "\u2029\u2029\u2029\u2029"
+    @test normalize("\u00f1", stripmark=true) == "n"
+    @test isempty(normalize("\u00ad", stripignore=true))
+    @test normalize("\t\r", stripcc=true) == "  "
+    @test normalize("\t\r", stripcc=true, newline2ls=true) == " \u2028"
 end
 
 @testset "unicode sa#15" begin
@@ -27,7 +31,7 @@ end
     #http://www.unicode.org/reports/tr15/
 
     @testset "canonical equivalence" begin
-        let ==(a::Array{Char},b::Array{Char}) = normalize_string(string(a...), :NFC)==normalize_string(string(b...), :NFC)
+        let ==(a::Array{Char},b::Array{Char}) = normalize(string(a...), :NFC)==normalize(string(b...), :NFC)
             ==(a,b) = Base.:(==)(a,b)
             @test ['C', '̧'] == ['Ç']
             @test ['q', '̇', '̣'] == ['q', '̣', '̇']
@@ -37,7 +41,7 @@ end
     end
 
     @testset "compatibility equivalence" begin
-        let ==(a::Array{Char},b::Array{Char}) = normalize_string(string(a...), :NFKC)==normalize_string(string(b...), :NFKC)
+        let ==(a::Array{Char},b::Array{Char}) = normalize(string(a...), :NFKC)==normalize(string(b...), :NFKC)
             ==(a,b) = Base.:(==)(a,b)
             @test ['ℌ'] == ['ℍ'] == ['H']
             @test ['ﻨ'] == ['ﻧ'] == ['ﻦ'] == ['ﻥ']
@@ -52,36 +56,36 @@ end
     end
 
     @testset "singletons" begin
-        @test normalize_string("\U212b", :NFD) == "A\U030a"
-        @test normalize_string("\U212b", :NFC) == "\U00c5"
-        @test normalize_string("\U2126", :NFC) == normalize_string("\U2126", :NFD) == "\U03a9"
+        @test normalize("\U212b", :NFD) == "A\U030a"
+        @test normalize("\U212b", :NFC) == "\U00c5"
+        @test normalize("\U2126", :NFC) == normalize("\U2126", :NFD) == "\U03a9"
     end
 
     @testset "canonical composites" begin
-        @test normalize_string("\U00c5", :NFC) == "\U00c5"
-        @test normalize_string("\U00c5", :NFD) == "A\U030a"
-        @test normalize_string("\U00f4", :NFC) == "\U00f4"
-        @test normalize_string("\U00f4", :NFD) == "o\U0302"
+        @test normalize("\U00c5", :NFC) == "\U00c5"
+        @test normalize("\U00c5", :NFD) == "A\U030a"
+        @test normalize("\U00f4", :NFC) == "\U00f4"
+        @test normalize("\U00f4", :NFD) == "o\U0302"
     end
 
     @testset "multiple combining marks" begin
-        @test normalize_string("\U1e69", :NFD) == "s\U0323\U0307"
-        @test normalize_string("\U1e69", :NFC) == "\U1e69"
-        @test normalize_string("\U1e0b\U0323", :NFD) == "d\U0323\U0307"
-        @test normalize_string("\U1e0b\U0323", :NFC) == "\U1e0d\U0307"
-        @test normalize_string("q\U0307\U0323", :NFC) == "q\U0323\U0307"
-        @test normalize_string("q\U0307\U0323", :NFD) == "q\U0323\U0307"
+        @test normalize("\U1e69", :NFD) == "s\U0323\U0307"
+        @test normalize("\U1e69", :NFC) == "\U1e69"
+        @test normalize("\U1e0b\U0323", :NFD) == "d\U0323\U0307"
+        @test normalize("\U1e0b\U0323", :NFC) == "\U1e0d\U0307"
+        @test normalize("q\U0307\U0323", :NFC) == "q\U0323\U0307"
+        @test normalize("q\U0307\U0323", :NFD) == "q\U0323\U0307"
     end
 
     @testset "compatibility composites" begin
-        @test normalize_string("\Ufb01", :NFD) == normalize_string("\Ufb01", :NFC) == "\Ufb01"
-        @test normalize_string("\Ufb01", :NFKD) == normalize_string("\Ufb01", :NFKC) == "fi"
-        @test normalize_string("2\U2075", :NFD) == normalize_string("2\U2075", :NFC) == "2\U2075"
-        @test normalize_string("2\U2075", :NFKD) == normalize_string("2\U2075", :NFKC) == "25"
-        @test normalize_string("\U1e9b\U0323", :NFD) == "\U017f\U0323\U0307"
-        @test normalize_string("\U1e9b\U0323", :NFC) == "\U1e9b\U0323"
-        @test normalize_string("\U1e9b\U0323", :NFKD) == "s\U0323\U0307"
-        @test normalize_string("\U1e9b\U0323", :NFKC) == "\U1e69"
+        @test normalize("\Ufb01", :NFD) == normalize("\Ufb01", :NFC) == "\Ufb01"
+        @test normalize("\Ufb01", :NFKD) == normalize("\Ufb01", :NFKC) == "fi"
+        @test normalize("2\U2075", :NFD) == normalize("2\U2075", :NFC) == "2\U2075"
+        @test normalize("2\U2075", :NFKD) == normalize("2\U2075", :NFKC) == "25"
+        @test normalize("\U1e9b\U0323", :NFD) == "\U017f\U0323\U0307"
+        @test normalize("\U1e9b\U0323", :NFC) == "\U1e9b\U0323"
+        @test normalize("\U1e9b\U0323", :NFKD) == "s\U0323\U0307"
+        @test normalize("\U1e9b\U0323", :NFKC) == "\U1e69"
     end
 end
 
@@ -92,7 +96,7 @@ end
         @test islower(c) == true
         @test isupper(c) == false
         @test isdigit(c) == false
-        @test isnumber(c) == false
+        @test isnumeric(c) == false
     end
 
     aupper=['A', 'D', 'J', 'Y', 'Z']
@@ -102,7 +106,7 @@ end
         @test islower(c) == false
         @test isupper(c) == true
         @test isdigit(c) == false
-        @test isnumber(c) == false
+        @test isnumeric(c) == false
     end
 
     nocase=['א','ﺵ']
@@ -110,7 +114,7 @@ end
 
     for c in alphas
         @test isalpha(c) == true
-        @test isnumber(c) == false
+        @test isnumeric(c) == false
     end
 
     anumber=['0', '1', '5', '9']
@@ -118,11 +122,11 @@ end
 
     for c in anumber
         @test isdigit(c) == true
-        @test isnumber(c) == true
+        @test isnumeric(c) == true
     end
     for c in unumber
         @test isdigit(c) == false
-        @test isnumber(c) == true
+        @test isnumeric(c) == true
     end
 
     alnums=vcat(alphas,anumber,unumber)
@@ -197,7 +201,7 @@ end
     @test !all(isgraph,"  \t   \n   \r  ")
     @test !all(isprint,"  \t   \n   \r  ")
     @test !all(isalpha,"  \t   \n   \r  ")
-    @test !all(isnumber,"  \t   \n   \r  ")
+    @test !all(isnumeric,"  \t   \n   \r  ")
     @test !all(ispunct,"  \t   \n   \r  ")
 
     @test !all(isspace,"ΣβΣβ")
@@ -206,11 +210,11 @@ end
     @test  all(isprint,"ΣβΣβ")
     @test !all(isupper,"ΣβΣβ")
     @test !all(islower,"ΣβΣβ")
-    @test !all(isnumber,"ΣβΣβ")
+    @test !all(isnumeric,"ΣβΣβ")
     @test !all(iscntrl,"ΣβΣβ")
     @test !all(ispunct,"ΣβΣβ")
 
-    @test  all(isnumber,"23435")
+    @test  all(isnumeric,"23435")
     @test  all(isdigit,"23435")
     @test  all(isalnum,"23435")
     @test !all(isalpha,"23435")
@@ -226,9 +230,9 @@ end
 @testset "utf8proc" begin
     # check utf8proc handling of CN category constants
     let c_ll = 'β', c_cn = '\u038B'
-        @test Base.UTF8proc.category_code(c_ll) == Base.UTF8proc.UTF8PROC_CATEGORY_LL
+        @test Base.Unicode.category_code(c_ll) == Base.Unicode.UTF8PROC_CATEGORY_LL
         # check codepoint with category code CN
-        @test Base.UTF8proc.category_code(c_cn) == Base.UTF8proc.UTF8PROC_CATEGORY_CN
+        @test Base.Unicode.category_code(c_cn) == Base.Unicode.UTF8PROC_CATEGORY_CN
     end
 end
 
@@ -246,8 +250,8 @@ end
         for T in (String,GenericString)
             for nf in (:NFC, :NFD)
                 for (s, g) in grphtest
-                    s_ = T(normalize_string(s, nf))
-                    g_ = map(s -> normalize_string(s, nf), g)
+                    s_ = T(normalize(s, nf))
+                    g_ = map(s -> normalize(s, nf), g)
                     # #9261
                     if length(s_) > 0
                         @test typeof(first(graphemes(s_))) == SubString{typeof(s_)}
@@ -257,7 +261,7 @@ end
                     @test grph == g_
                     @test length(graphemes(s_)) == length(grph)
                 end
-                S = [T(normalize_string(s)) for (s,g) in grphtest]
+                S = [T(normalize(s)) for (s,g) in grphtest]
                 G = map(graphemes, S)
                 @test map(graphemes, sort!(S)) == sort!(G)
             end
@@ -277,23 +281,23 @@ end
 @testset "#10958 handling of embedded NUL chars" begin
     @test length("\0w") == length("\0α") == 2
     @test textwidth("\0w") == textwidth("\0α") == 1
-    @test normalize_string("\0W", casefold=true) == "\0w"
+    @test normalize("\0W", casefold=true) == "\0w"
 end
 
 @testset "ut8proc_map with GenericString" begin
-    @test normalize_string(GenericString("\u006e\u0303"), :NFC) == "\u00f1"
+    @test normalize(GenericString("\u006e\u0303"), :NFC) == "\u00f1"
 end
 
-@testset "normalize_string keywords" begin
-    @test_throws ArgumentError normalize_string("\u006e\u0303", compose=false, compat=true)
-    @test_throws ArgumentError normalize_string("\u006e\u0303", compose=false, stripmark=true)
+@testset "normalize keywords" begin
+    @test_throws ArgumentError normalize("\u006e\u0303", compose=false, compat=true)
+    @test_throws ArgumentError normalize("\u006e\u0303", compose=false, stripmark=true)
 end
 
 @testset "fastplus" begin
     @test lowercase('A') == 'a'
     @test uppercase('a') == 'A'
 
-    @test is_assigned_char('A')
+    @test isassigned('A')
 end
 
 @testset "isspace" begin
@@ -319,4 +323,73 @@ end
     g = graphemes(SubString("123α56789", 1, 6))
     @test eltype(g) == SubString{String}
     @test collect(g) == ["1","2","3","α","5"]
+end
+
+@testset "ucfirst/lcfirst" begin
+    @test ucfirst("Hola")=="Hola"
+    @test ucfirst("hola")=="Hola"
+    @test ucfirst("")==""
+    @test ucfirst("*")=="*"
+    @test ucfirst("Ǆxx") == ucfirst("ǆxx") == "ǅxx"
+
+    @test lcfirst("Hola")=="hola"
+    @test lcfirst("hola")=="hola"
+    @test lcfirst("")==""
+    @test lcfirst("*")=="*"
+end
+
+@testset "issue #11482" begin
+    @testset "uppercase/lowercase" begin
+        @test uppercase("aBc") == "ABC"
+        @test uppercase('A') == 'A'
+        @test uppercase('a') == 'A'
+        @test lowercase("AbC") == "abc"
+        @test lowercase('A') == 'a'
+        @test lowercase('a') == 'a'
+        @test uppercase('α') == '\u0391'
+        @test lowercase('Δ') == 'δ'
+        @test lowercase('\U118bf') == '\U118df'
+        @test uppercase('\U1044d') == '\U10425'
+    end
+    @testset "ucfirst/lcfirst" begin
+        @test ucfirst("Abc") == "Abc"
+        @test ucfirst("abc") == "Abc"
+        @test lcfirst("ABC") == "aBC"
+        @test lcfirst("aBC") == "aBC"
+        @test ucfirst(GenericString("")) == ""
+        @test lcfirst(GenericString("")) == ""
+        @test ucfirst(GenericString("a")) == "A"
+        @test lcfirst(GenericString("A")) == "a"
+        @test lcfirst(GenericString("a")) == "a"
+        @test ucfirst(GenericString("A")) == "A"
+    end
+    @testset "titlecase" begin
+        @test titlecase('ǉ') == 'ǈ'
+        @test titlecase("ǉubljana") == "ǈubljana"
+        @test titlecase("aBc ABC") == "ABc ABC"
+        @test titlecase("abcD   EFG\n\thij") == "AbcD   EFG\n\tHij"
+    end
+end
+
+@testset "issue # 11464: uppercase/lowercase of GenericString becomes a String" begin
+    str = "abcdef\uff\uffff\u10ffffABCDEF"
+    @test typeof(uppercase("abcdef")) == String
+    @test typeof(uppercase(GenericString(str))) == String
+    @test typeof(lowercase("ABCDEF")) == String
+    @test typeof(lowercase(GenericString(str))) == String
+
+    foomap(ch) = (ch > Char(65))
+    foobar(ch) = Char(0xd800)
+    foobaz(ch) = reinterpret(Char, typemax(UInt32))
+    @test_throws ArgumentError map(foomap, GenericString(str))
+    @test map(foobar, GenericString(str)) == String(repeat(b"\ud800", outer=[17]))
+    @test map(foobaz, GenericString(str)) == String(repeat(b"\ufffd", outer=[17]))
+
+    @test "a".*["b","c"] == ["ab","ac"]
+    @test ["b","c"].*"a" == ["ba","ca"]
+    @test ["a","b"].*["c" "d"] == ["ac" "ad"; "bc" "bd"]
+
+    @test one(String) == ""
+    @test prod(["*" for i in 1:3]) == "***"
+    @test prod(["*" for i in 1:0]) == ""
 end